refactor: consolidate organization slug generation into OrganizationService and update sign-up process
