+++ web-app/src/lib/services/organization.service.ts
@@ -1,3 +1,6 @@
+import { nanoid } from "nanoid";
+import slugify from "slugify";
+
 import {
   organizationInclude,
   organizationOrderBy,
@@ -107,5 +110,34 @@ export class OrganizationService extends BaseService<OrganizationService> {
       data,
       include: organizationInclude,
     });
+  }
+
+  /**
+   * Generates a unique, URL-friendly slug for an organization based on its name.
+   *
+   * - Converts the provided name to a slug using lowercase and strict mode.
+   * - Checks if an organization with the generated slug already exists.
+   *   - If not, returns the slug.
+   *   - If it exists, appends a unique 6-character ID to ensure uniqueness.
+   *
+   * @param name - The name of the organization to generate a slug for.
+   * @returns A Promise resolving to a unique slug string.
+   *
+   * @example
+   * ```typescript
+   * const slug = await service.generateOrganizationSlugFromName("Acme Inc");
+   * // Possible result: "acme-inc" or "acme-inc-1a2b3c"
+   * ```
+   */
+  async generateOrganizationSlugFromName(name: string): Promise<string> {
+    const slugedName = slugify(name, { lower: true, strict: true });
+    const existingOrganization =
+      await OrganizationService.getInstance().getOrganizationBySlug(slugedName);
+    if (!existingOrganization) {
+      return slugedName;
+    }
+
+    const uniqueId = nanoid(6);
+    return `${slugedName}-${uniqueId}`;
   }
 }
